+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # NEXT RELEASE
 * Rename `TheAlgorithm.extractSummaryInfo()` to `TheAlgorithm.initializeFiltersWithSummaryInfo()`
+* Allow things that user has reblogged into the feed
+* `Toot.realAccount()` method
 
 # v0.19.0
 * Preserve `MastodonServersInfo` information like MAU